Moved most of stregister binary code to lib.
